реализует отображение main, когда нет предложений
